+++ cjs-entry.d.ts
@@ -1,3 +1,0 @@
-declare const _default: typeof import('./lib/cjs/puppeteer/node').default;
-
-export default _default;