+++ src/common/BrowserConnector.ts
@@ -24,12 +24,24 @@ import { Viewport } from './PuppeteerViewport.js';
 import { isNode } from '../environment.js';
 
 /**
- * Generic browser options that can be passed when launching any browser.
+ * Generic browser options that can be passed when launching any browser or when
+ * connecting to an existing browser instance.
  * @public
  */
-export interface BrowserOptions {
+export interface BrowserConnectOptions {
+  /**
+   * Whether to ignore HTTPS errors during navigation.
+   * @defaultValue false
+   */
   ignoreHTTPSErrors?: boolean;
+  /**
+   * Sets the viewport for each page.
+   */
   defaultViewport?: Viewport;
+  /**
+   * Slows down Puppeteer operations by the specified amount of milliseconds to
+   * aid debugging.
+   */
   slowMo?: number;
 }
 
@@ -46,7 +58,7 @@ const getWebSocketTransportClass = async () => {
  * @internal
  */
 export const connectToBrowser = async (
-  options: BrowserOptions & {
+  options: BrowserConnectOptions & {
     browserWSEndpoint?: string;
     browserURL?: string;
     transport?: ConnectionTransport;