+++ src/common/NetworkConditions.ts
@@ -16,8 +16,14 @@
 
 import { NetworkConditions } from './NetworkManager.js';
 
+/**
+ * @public
+ */
 export type PredefinedNetworkConditions = { [name: string]: NetworkConditions };
 
+/**
+ * @public
+ */
 export const networkConditions: PredefinedNetworkConditions = {
   'Slow 3G': {
     download: ((500 * 1000) / 8) * 0.8,